[feat](shotgun) Split shotgun script in different classes
